Pipeline: using a parent class for plotdata plugins
