+++ pydatview/common.py
@@ -444,8 +444,9 @@ def isDate(x):
 # Create a Dummy Main Frame Class for testing purposes (e.g. of plugins)
 
 class DummyMainFrame():
-    def __init__(self, parent): self.parent=parent
+    def __init__(self, parent): self.parent=parent; 
     def addAction            (self, *args, **kwargs): Info(self.parent, 'This is dummy '+inspect.stack()[0][3])
     def removeAction         (self, *args, **kwargs): Info(self.parent, 'This is dummy '+inspect.stack()[0][3])
     def load_dfs             (self, *args, **kwargs): Info(self.parent, 'This is dummy '+inspect.stack()[0][3])
     def mainFrameUpdateLayout(self, *args, **kwargs): Info(self.parent, 'This is dummy '+inspect.stack()[0][3])
+    def redraw               (self, *args, **kwargs): Info(self.parent, 'This is dummy '+inspect.stack()[0][3])
