+++ pydatview/plugins/plotdata_removeOutliers.py
@@ -1,9 +1,8 @@
 import wx
 import numpy as np
-from pydatview.GUITools import GUIToolPanel, TOOL_BORDER
-from pydatview.common import CHAR, Error, Info, pretty_num_short
-from pydatview.common import DummyMainFrame
-from pydatview.plotdata import PlotData
+from pydatview.GUITools import TOOL_BORDER
+from pydatview.plugins.plotdata_default_plugin import PlotDataActionEditor
+from pydatview.common import Error, Info
 from pydatview.pipeline import PlotDataAction
 import platform
 # --------------------------------------------------------------------------------}
@@ -55,28 +54,16 @@ def removeOutliersXY(x, y, opts):
 # --------------------------------------------------------------------------------}
 # --- GUI to edit plugin and control the action
 # --------------------------------------------------------------------------------{
-class RemoveOutliersToolPanel(GUIToolPanel):
+class RemoveOutliersToolPanel(PlotDataActionEditor):
 
-    def __init__(self, parent, action):
-        GUIToolPanel.__init__(self, parent)
+    def __init__(self, parent, action, plotPanel, pipeLike, **kwargs):
+        PlotDataActionEditor.__init__(self, parent, action, plotPanel, pipeLike, tables=False, buttons=[''])
 
-        # --- Creating "Fake data" for testing only!
-        if action is None:
-            print('[WARN] Calling GUI without an action! Creating one.')
-            mainframe = DummyMainFrame(parent)
-            action = binningAction(label='dummyAction', mainframe=mainframe)
-
-        # --- Data
-        self.parent = parent # parent is GUIPlotPanel
-        self.mainframe = action.mainframe
-        self.data = action.data
-        self.action = action
         # --- GUI elements
-        self.btClose = self.getBtBitmap(self,'Close','close',self.destroy)
-        self.btApply = self.getToggleBtBitmap(self,'Apply','cloud',self.onToggleApply)
+        #self.btClose = self.getBtBitmap(self,'Close','close',self.destroy)
+        #self.btApply = self.getToggleBtBitmap(self,'Apply','cloud',self.onToggleApply)
 
         lb1 = wx.StaticText(self, -1, 'Median deviation:')
-#         self.tMD = wx.TextCtrl(self, wx.ID_ANY,, size = (30,-1), style=wx.TE_PROCESS_ENTER)
         self.tMD = wx.SpinCtrlDouble(self, value='11', size=wx.Size(60,-1))
         self.tMD.SetValue(5)
         self.tMD.SetRange(0.0, 1000)
@@ -85,12 +72,14 @@ class RemoveOutliersToolPanel(GUIToolPanel):
         self.lb = wx.StaticText( self, -1, '')
         
         # --- Layout        
-        self.sizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.sizer.Add(self.btClose,0,flag = wx.LEFT|wx.CENTER,border = 1)
-        self.sizer.Add(self.btApply,0,flag = wx.LEFT|wx.CENTER,border = 5)
-        self.sizer.Add(lb1         ,0,flag = wx.LEFT|wx.CENTER,border = 5)
-        self.sizer.Add(self.tMD    ,0,flag = wx.LEFT|wx.CENTER,border = 5)
-        self.sizer.Add(self.lb     ,0,flag = wx.LEFT|wx.CENTER,border = 5)
+        hsizer = wx.BoxSizer(wx.HORIZONTAL)
+#         self.sizer.Add(self.btClose,0,flag = wx.LEFT|wx.CENTER,border = 1)
+#         self.sizer.Add(self.btApply,0,flag = wx.LEFT|wx.CENTER,border = 5)
+        hsizer.Add(lb1         ,0,flag = wx.LEFT|wx.CENTER,border = 5)
+        hsizer.Add(self.tMD    ,0,flag = wx.LEFT|wx.CENTER,border = 5)
+        hsizer.Add(self.lb     ,0,flag = wx.LEFT|wx.CENTER,border = 5)
+
+        self.sizer.Add(hsizer  ,0,flag = wx.LEFT|wx.CENTER,border = 5)
         self.SetSizer(self.sizer)
 
         # --- Events
@@ -107,40 +96,40 @@ class RemoveOutliersToolPanel(GUIToolPanel):
         self._Data2GUI()        
         self.onToggleApply(init=True)
 
-    # --- Implementation specific
-
     # --- Bindings for plot triggers on parameters changes
-    def onParamChange(self, event=None):
-        self._GUI2Data()
-        if self.data['active']:
-            self.parent.load_and_draw() # Data will change
-
-    def onParamChangeArrow(self, event):
-        self.onParamChange()
-        event.Skip()
-
-    def onParamChangeEnter(self, event):
-        self.onParamChange()
-        event.Skip()
-
-    def onParamChangeChar(self, event):
-        event.Skip()  
-        code = event.GetKeyCode()
-        if code in [wx.WXK_RETURN, wx.WXK_NUMPAD_ENTER]:
-            #print(self.spintxt.Value)
-            self.tMD.SetValue(self.spintxt.Value)
-            self.onParamChangeEnter(event)
+#     def onParamChange(self, event=None):
+#         self._GUI2Data()
+#         if self.data['active']:
+#             self.parent.load_and_draw() # Data will change
+# 
+#     def onParamChangeArrow(self, event):
+#         self.onParamChange()
+#         event.Skip()
+# 
+#     def onParamChangeEnter(self, event):
+#         self.onParamChange()
+#         event.Skip()
+# 
+#     def onParamChangeChar(self, event):
+#         event.Skip()  
+#         code = event.GetKeyCode()
+#         if code in [wx.WXK_RETURN, wx.WXK_NUMPAD_ENTER]:
+#             #print(self.spintxt.Value)
+#             self.tMD.SetValue(self.spintxt.Value)
+#             self.onParamChangeEnter(event)
 
     # --- Table related
     # --- External Calls
-    def cancelAction(self, redraw=True):
+    def cancelAction(self):
         """ do cancel the action"""
         self.lb.SetLabel('Click on "Apply" to remove outliers on the fly for all new plot.')
-        self.btApply.SetLabel(CHAR['cloud']+' Apply')
-        self.btApply.SetValue(False)
-        self.data['active'] = False     
-        if redraw:
-            self.parent.load_and_draw() # Data will change based on plotData 
+        PlotDataActionEditor.cancelAction(self)
+
+#         self.btApply.SetLabel(CHAR['cloud']+' Apply')
+#         self.btApply.SetValue(False)
+#         self.data['active'] = False     
+#         if redraw:
+#             self.parent.load_and_draw() # Data will change based on plotData 
 
     # --- Fairly generic
     def _GUI2Data(self):
@@ -148,26 +137,30 @@ class RemoveOutliersToolPanel(GUIToolPanel):
 
     def _Data2GUI(self):
         self.tMD.SetValue(self.data['medianDeviation'])
+# 
+#     def onToggleApply(self, event=None, init=False):
+# 
+#         if not init:
+#             self.data['active'] = not self.data['active']
+# 
+#         if self.data['active']:
+#             self._GUI2Data()
+#             self.lb.SetLabel('Outliers are now removed on the fly. Click "Clear" to stop.')
+#             self.btApply.SetLabel(CHAR['sun']+' Clear')
+#             self.btApply.SetValue(True)            
+#             # The action is now active we add it to the pipeline, unless it's already in it
+#             if self.mainframe is not None:
+#                 self.mainframe.addAction(self.action, overwrite=True)
+#             if not init:
+#                 self.parent.load_and_draw() # filter will be applied in plotData.py
+#         else:
+#             # We remove our action from the pipeline
+#             if not init:
+#                 if self.mainframe is not None:
+#                     self.mainframe.removeAction(self.action)          
+#             self.cancelAction(redraw= not init)
 
-    def onToggleApply(self, event=None, init=False):
+if __name__ == '__main__':
+    from pydatview.plugins.plotdata_default_plugin import demoPlotDataActionPanel
 
-        if not init:
-            self.data['active'] = not self.data['active']
-
-        if self.data['active']:
-            self._GUI2Data()
-            self.lb.SetLabel('Outliers are now removed on the fly. Click "Clear" to stop.')
-            self.btApply.SetLabel(CHAR['sun']+' Clear')
-            self.btApply.SetValue(True)            
-            # The action is now active we add it to the pipeline, unless it's already in it
-            if self.mainframe is not None:
-                self.mainframe.addAction(self.action, overwrite=True)
-            if not init:
-                self.parent.load_and_draw() # filter will be applied in plotData.py
-        else:
-            # We remove our action from the pipeline
-            if not init:
-                if self.mainframe is not None:
-                    self.mainframe.removeAction(self.action)          
-            self.cancelAction(redraw= not init)
-
+    demoPlotDataActionPanel(RemoveOutliersToolPanel, plotDataFunction=removeOutliersXY, data=_DEFAULT_DICT)
